+++ angular-workspace/projects/ni/nimble-angular/src/directives/table-column/text/nimble-table-column-text.directive.ts
@@ -1,5 +1,6 @@
 import { Directive, ElementRef, Input, Renderer2 } from '@angular/core';
 import type { TableColumnText } from '@ni/nimble-components/dist/esm/table-column/text';
+import { NimbleTableColumnBaseDirective } from '../base/nimble-table-column-base.directive';
 import type { NumberValueOrAttribute } from '../../utilities/template-value-helpers';
 
 export type { TableColumnText };
@@ -10,17 +11,7 @@ export type { TableColumnText };
 @Directive({
     selector: 'nimble-table-column-text'
 })
-export class NimbleTableColumnTextDirective {
-    public get columnId(): string | undefined {
-        return this.elementRef.nativeElement.columnId;
-    }
-
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
-    @Input('column-id') public set columnId(value: string | undefined) {
-        this.renderer.setProperty(this.elementRef.nativeElement, 'columnId', value);
-    }
-
+export class NimbleTableColumnTextDirective extends NimbleTableColumnBaseDirective<TableColumnText> {
     public get fieldName(): string | undefined {
         return this.elementRef.nativeElement.fieldName;
     }
@@ -59,5 +50,7 @@ export class NimbleTableColumnTextDirective {
         this.renderer.setProperty(this.elementRef.nativeElement, 'minPixelWidth', value);
     }
 
-    public constructor(private readonly renderer: Renderer2, private readonly elementRef: ElementRef<TableColumnText>) {}
+    public constructor(renderer: Renderer2, elementRef: ElementRef<TableColumnText>) {
+        super(renderer, elementRef);
+    }
 }