+++ angular-workspace/projects/ni/nimble-angular/src/directives/table/nimble-table.directive.ts
@@ -1,9 +1,10 @@
 import { Directive, ElementRef, Input, OnDestroy, Renderer2 } from '@angular/core';
 import type { Table } from '@ni/nimble-components/dist/esm/table';
-import type { TableRecord, TableFieldName, TableFieldValue, TableValidity } from '@ni/nimble-components/dist/esm/table/types';
+import type { TableRecord, TableFieldName, TableFieldValue, TableValidity, TableActionMenuToggleEventDetail } from '@ni/nimble-components/dist/esm/table/types';
 import type { Observable, Subscription } from 'rxjs';
 
 export type { Table };
+export type { TableActionMenuToggleEventDetail };
 export { TableRecord, TableFieldName, TableFieldValue, TableValidity };
 
 /**