+++ packages/nimble-components/src/table-column/text/table-column-text-base.ts
@@ -1,5 +1,5 @@
 import { attr } from '@microsoft/fast-element';
-import type { TableStringField, TableFieldName } from '../../table/types';
+import type { TableStringField } from '../../table/types';
 import { TableColumn } from '../base';
 import { cellStyles } from './styles';
 import { cellTemplate } from './template';
@@ -28,11 +28,11 @@ TableColumnTextColumnConfig
 
     public readonly cellTemplate = cellTemplate;
 
-    public getColumnConfig(): TableColumnTextColumnConfig {
-        return { placeholder: this.placeholder ?? '' };
+    protected fieldNameChanged(): void {
+        this.dataRecordFieldNames = [this.fieldName] as const;
     }
 
-    public getDataRecordFieldNames(): (TableFieldName | undefined)[] {
-        return [this.fieldName];
+    protected placeholderChanged(): void {
+        this.columnConfig = { placeholder: this.placeholder ?? '' };
     }
 }