+++ packages/nimble-components/src/table/template.ts
@@ -12,6 +12,7 @@ import type { Table } from '.';
 import { TableHeader } from './components/header';
 import { TableRow } from './components/row';
 import type { TableColumn } from '../table-column/base';
+import type { TableActionMenuToggleEventDetail } from './types';
 
 // prettier-ignore
 export const template = html<Table>`
@@ -20,16 +21,18 @@ export const template = html<Table>`
             <div role="rowgroup" class="header-container">
                 <div class="header-row" role="row" style="grid-template-columns: ${x => x.rowGridColumns} auto; left: -${x => x.scrollX}px;">
                     ${repeat(x => x.columns, html<TableColumn>`
-                        <${DesignSystem.tagFor(TableHeader)} class="header">
-                            <slot name="${x => x.slot}"></slot>
-                        </${DesignSystem.tagFor(TableHeader)}>
+                        ${when(x => !x.columnHidden, html<TableColumn>`
+                            <${DesignSystem.tagFor(TableHeader)} class="header">
+                                <slot name="${x => x.slot}"></slot>
+                            </${DesignSystem.tagFor(TableHeader)}>
+                        `)}
                     `)}
                     <div class="header-scrollbar-spacer" style="width: ${x => (x.virtualizer.headerContainerMarginRight !== 0 ? '17' : '0')}px;"></div>
                 </div>
             </div>
             <div class="table-viewport" ${ref('viewport')}>
                 <div class="table-scroll" style="height: ${x => x.virtualizer.allRowsHeight}px;"></div>
-                <div class="table-row-container" role="rowgroup" style="transform: ${x => (x.virtualizer.rowContainerYOffset === 0 ? 'none' : `translateY(${x.virtualizer.rowContainerYOffset}px)`)};">
+                <div class="table-row-container" role="rowgroup" style="top: ${x => `${x.virtualizer.rowContainerYOffset}px;`}">
                     ${when(x => x.columns.length > 0 && x.canRenderRows, html<Table>`
                         ${repeat(x => x.virtualizer.visibleItems, html<VirtualItem, Table>`
                             <${DesignSystem.tagFor(TableRow)}
@@ -38,7 +41,17 @@ export const template = html<Table>`
                                 :dataRecord="${(x, c) => c.parent.tableData[x.index]?.record}"
                                 :columns="${(_, c) => c.parent.columns}"
                                 :rowGridColumns="${(_, c) => c.parent.rowGridColumns}"
+                                @row-action-menu-beforetoggle="${(_, c) => c.parent.onRowActionMenuBeforeToggle(c.event as CustomEvent<TableActionMenuToggleEventDetail>)}"
+                                @row-action-menu-toggle="${(_, c) => c.parent.onRowActionMenuToggle(c.event as CustomEvent<TableActionMenuToggleEventDetail>)}"
                             >
+                            ${when((x, c) => (c.parent as Table).openActionMenuRecordId === (c.parent as Table).tableData[x.index]?.id, html<VirtualItem, Table>`
+                                ${repeat((_, c) => (c.parent as Table).actionMenuSlots, html<string, Table>`
+                                    <slot
+                                        name="${x => x}"
+                                        slot="${x => `row-action-menu-${x}`}">
+                                    </slot>
+                                `)}
+                            `)}                        
                             </${DesignSystem.tagFor(TableRow)}>
                         `)}
                     `)}