+++ packages/nimble-components/src/table/tests/table.pageobject.ts
@@ -2,6 +2,7 @@ import type { Table } from '..';
 import type { TableHeader } from '../components/header';
 import type { TableRecord } from '../types';
 import { waitForUpdatesAsync } from '../../testing/async-helpers';
+import type { MenuButton } from '../../menu-button';
 
 /**
  * Page object for the `nimble-table` component to provide consistent ways
@@ -15,6 +16,19 @@ export class TablePageObject<T extends TableRecord> {
             'nimble-table-header'
         )!;
         return headers.length;
+    }
+
+    public getRenderedCellCountForRow(rowIndex: number): number {
+        const rows = this.tableElement.shadowRoot!.querySelectorAll('nimble-table-row');
+        if (rowIndex >= rows.length) {
+            throw new Error(
+                'Attempting to index past the total number of rendered rows'
+            );
+        }
+
+        const row = rows.item(rowIndex);
+        const cells = row.shadowRoot!.querySelectorAll('nimble-table-cell');
+        return cells.length;
     }
 
     public getHeaderContent(columnIndex: number): Node | undefined {
@@ -156,6 +170,80 @@ export class TablePageObject<T extends TableRecord> {
         const scrollElement = this.tableElement.viewport;
         scrollElement.scrollTop = scrollElement.scrollHeight;
         await waitForUpdatesAsync();
+    }
+
+    public getCellActionMenu(
+        rowIndex: number,
+        columnIndex: number
+    ): MenuButton | null {
+        const rows = this.tableElement.shadowRoot!.querySelectorAll('nimble-table-row');
+        if (rowIndex >= rows.length) {
+            throw new Error(
+                'Attempting to index past the total number of rendered rows'
+            );
+        }
+
+        const row = rows.item(rowIndex);
+        const cells = row.shadowRoot!.querySelectorAll('nimble-table-cell');
+        if (columnIndex >= cells.length) {
+            throw new Error(
+                'Attempting to index past the total number of rendered columns'
+            );
+        }
+
+        return cells
+            .item(columnIndex)
+            .shadowRoot!.querySelector<MenuButton>('nimble-menu-button');
+    }
+
+    public async clickCellActionMenu(
+        rowIndex: number,
+        columnIndex: number
+    ): Promise<void> {
+        this.setRowHoverState(rowIndex, true);
+        await waitForUpdatesAsync();
+
+        const menuButton = this.getCellActionMenu(rowIndex, columnIndex);
+        if (!menuButton) {
+            throw new Error('Cannot click on a non-visible action menu');
+        }
+
+        menuButton.toggleButton!.control.click();
+    }
+
+    public isCellActionMenuVisible(
+        rowIndex: number,
+        columnIndex: number
+    ): boolean {
+        const actionMenu = this.getCellActionMenu(rowIndex, columnIndex);
+        if (!actionMenu) {
+            return false;
+        }
+
+        return window.getComputedStyle(actionMenu).display !== 'none';
+    }
+
+    public setRowHoverState(rowIndex: number, hover: boolean): void {
+        const rows = this.tableElement.shadowRoot!.querySelectorAll('nimble-table-row');
+        if (rowIndex >= rows.length) {
+            throw new Error(
+                'Attempting to index past the total number of rendered rows'
+            );
+        }
+
+        const cells = rows
+            .item(rowIndex)
+            .shadowRoot!.querySelectorAll('nimble-table-cell');
+        if (hover) {
+            cells.forEach(cell => cell.style.setProperty(
+                '--ni-private-table-cell-action-menu-display',
+                'block'
+            ));
+        } else {
+            cells.forEach(cell => cell.style.removeProperty(
+                '--ni-private-table-cell-action-menu-display'
+            ));
+        }
     }
 
     private getHeaderContentElement(